introduce and use WaitForOnboarded
